pyproject.toml
[project]
name = "leonardo-ml-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "gilbertofp16",email = "gilbertofp16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "pillow>=10.4.0",
    "httpx>=0.27.0",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "tqdm>=4.66.4",
    "transformers>=4.42.4",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "text_image_similarity" },
]
